[!!!][TASK] Change namespace for classes and package name
Breaking Change!
For better publishing and preventing confusion the package
name was changed. Because of that I changed the namespace for the
classes too.
Pay attention to this and change classname in extending extension,
used viewhelper and so on.
I replaced recursively in extension folder with the following pattern

* kn/ -> kanow/
* KN/ -> Kanow/
* KN\ -> Kanow\

Settings for match case activated.

That should do the work for you.